Add type restrictions to supply parameters to protect agains multiplication overflow

+++ contracts/vending/VendingMachine.sol
@@ -65,20 +65,23 @@ contract VendingMachine is Ownable, IReceiveApproval {
         uint256 wrappedTokenAmount
     );
 
-    /// @dev Sets the reference to `wrappedToken` and `tToken`. Initializes
-    ///      conversion ratio between wrapped token and T based on the provided
-    ///      `_tTokenAllocation` and `_wrappedTokenSupply`.
+    /// @notice Sets the reference to `wrappedToken` and `tToken`. Initializes
+    ///         conversion `ratio` between wrapped token and T based on the
+    ///         provided `_tTokenAllocation` and `_wrappedTokenSupply`.
     /// @param _wrappedToken Address to ERC20 token that will be wrapped to T
     /// @param _tToken Address of T token
     /// @param _wrappedTokenSupply The total supply of the token that will be
     ///       wrapped to T
     /// @param _tTokenAllocation The allocation of T this instance of Vending
     ///        Machine will receive
+    /// @dev Multiplications in this contract can't overflow uint256 as we
+    ///     restrict `_wrappedTokenSupply` and `_tTokenAllocation` to 192 bits
+    ///     and the value in FLOATING_POINT_DIVISOR fits in less than 60 bits.
     constructor(
         IERC20 _wrappedToken,
         T _tToken,
-        uint256 _wrappedTokenSupply,
-        uint256 _tTokenAllocation
+        uint192 _wrappedTokenSupply,
+        uint192 _tTokenAllocation
     ) {
         wrappedToken = _wrappedToken;
         tToken = _tToken;
